requirements.txt
Django>=2.0,<2.1
psycopg2-binary>=2.8
django-jet
six
django-computed-property
WeasyPrint>=52,<53
django-input-mask
django-localflavor